Fix menu (and keyevent) handling on Mac
Mod command as meta in menues.

sdl mapped command as meta and ctrl as ctrl,
but wxWidgets maps command as ctrl and ctrl as rawctrl.

There might be more bugs in this area..

Delay menubar frame attach until after the menus are created
so that about and preferences goto the correct place